requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
Werkzeug==0.10
alembic==0.7.4
gunicorn==19.1.1
itsdangerous==0.24
py-trello==0.4.2
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
pg8000==1.10.1
